Extracting function, adding tests.
