+++ graylog2-web-interface/src/util/FormsUtils.js
@@ -1,4 +1,5 @@
 import NumberUtils from 'util/NumberUtils';
+import createEvent from './CreateEvent';
 
 const FormUtils = {
   getValueFromInput(input) {
@@ -16,22 +17,12 @@ const FormUtils = {
   triggerInput(urlInput) {
     const { input } = urlInput;
     const tracker = input._valueTracker;
-    const event = this.createEvent('change');
+    const event = createEvent('change');
     event.simulated = true;
     if (tracker) {
       tracker.setValue('');
     }
     input.dispatchEvent(event);
-  },
-  // Workaround for IE11, see #7670
-  createEvent(type) {
-    if (typeof (Event) === 'function') {
-      return new Event(type);
-    }
-
-    const event = document.createEvent('Event');
-    event.initEvent(type, true, true);
-    return event;
   },
 };
